# Performance improvement of `cupy.intersect1d`