--- a/cupy/_sorting/search.py
+++ b/cupy/_sorting/search.py
@@ -373,7 +373,7 @@ _exists_kernel = _core.ElementwiseKernel(
     + _searchsorted_code + '''
     out = (y == n_bins ? false : bins[y] == x);
     if (invert) out = !out;
-    ''', name='cupy_searchsorted_kernel', preamble=_preamble+_hip_preamble)
+    ''', name='cupy_exists_kernel', preamble=_preamble+_hip_preamble)
 
 
 _exists_and_searchsorted_kernel = _core.ElementwiseKernel(
@@ -386,7 +386,7 @@ _exists_and_searchsorted_kernel = _core.ElementwiseKernel(
     + _searchsorted_code + '''
     out = (y == n_bins ? false : bins[y] == x);
     if (invert) out = !out;
-    ''', name='cupy_searchsorted_kernel', preamble=_preamble+_hip_preamble)
+    ''', name='cupy_exists_and_searchsorted_kernel', preamble=_preamble+_hip_preamble)
 
 
 def searchsorted(a, v, side='left', sorter=None):