Rel. #6402

master:

```
intersect1d         :    CPU:34600.527 us   +/-14.160 (min:34568.779 / max:34668.362) us     GPU-0:35218.770 us   +/-13.787 (min:35187.714 / max:35282.944) us
intersect1d         :    CPU:72343.659 us   +/-71.981 (min:72294.853 / max:72748.705) us     GPU-0:74817.813 us   +/-59.652 (min:74770.432 / max:75154.434) us
```

This PR:

```
intersect1d         :    CPU:24760.175 us   +/-19.352 (min:24705.804 / max:24900.344) us     GPU-0:25380.608 us   +/-18.880 (min:25326.591 / max:25516.031) us
intersect1d         :    CPU:56950.079 us   +/-51.338 (min:56914.579 / max:57307.126) us     GPU-0:58647.603 us   +/-49.220 (min:58611.713 / max:58987.518) us
```

Benchmark:

```py
import cupy
from cupy import testing
from cupyx.profiler import benchmark


shape = (64, 1024, 1024)
x1 = testing.shaped_random(shape, xp=cupy, dtype='l', seed=0, scale=100000000)
x2 = testing.shaped_random(shape, xp=cupy, dtype='l', seed=1, scale=100000000)
perf = benchmark(cupy.intersect1d, (x1, x2), {'return_indices': False}, n_repeat=100)
print(perf)
perf = benchmark(cupy.intersect1d, (x1, x2), {'return_indices': True}, n_repeat=100)
print(perf)
```

cc: @khushi-411